Refactored stateful user messages into new file
